SWDEV-269439 - Fix dtests build failure on HIP CUDA


[ROCm/hip commit: 4f062fa1fbaff90c5a777f64f9ebe7f2e7f3b32e]

+++ tests/src/kernel/hipDynamicShared.cpp
@@ -74,7 +74,7 @@ __global__ void testExternSharedKernel(const T* A_d, const T* B_d, T* C_d,
 }
 
 template <typename T>
-void testExternShared(size_t N, size_t groupElements) {
+void testExternShared(size_t N, unsigned groupElements) {
     size_t Nbytes = N * sizeof(T);
 
     T *A_d, *B_d, *C_d;
